Move verify_archive_node() to eth_defi.provider.broken_provider and log HTTP response headers on failure
